fix dtype in decoder export

+++ scripts/export_onnx_model.py
@@ -133,7 +133,7 @@ def run_export(
         "point_labels": torch.randint(low=0, high=4, size=(1, 5), dtype=torch.float),
         "mask_input": torch.randn(1, 1, *mask_input_size, dtype=torch.float),
         "has_mask_input": torch.tensor([1], dtype=torch.float),
-        "orig_im_size": torch.tensor([1500, 2250], dtype=torch.float),
+        "orig_im_size": torch.tensor([1500, 2250], dtype=torch.int32),
     }
 
     _ = onnx_model(**dummy_inputs)
